big-a/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/big_a"]

[project]
name = "big-a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pyqlib>=0.9.7",
  "torch",
  "lightgbm>=4.0",
  "scikit-learn",
  "pandas",
  "numpy",
  "matplotlib",
  "pyyaml",
  "loguru",
  "pytest",
  "ruff",
  "jupyter",
  "nbconvert",
  "safetensors",
  "transformers",
  "huggingface-hub",
  "einops",
  "typer",
  "langchain>=0.3.7",
  "langgraph>=0.2.56",
  "langchain-openai>=0.3.5",
  "langchain-core>=0.3.0",
  "langchain-community>=0.3.0",
  "protobuf>=3.20,<4.0",
  "mlflow>=2.0",
  "plotly>=5.0",
  "pyarrow",
]

# pytest configuration (minimal, project-wide)
[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "skip_if_no_data: skip test if Qlib data is not available",
]
